% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{vector: cumulative customer arrival times in seconds}

\item{ServiceTimes}{vector: duration of service time in seconds, must be same length as arrival}

\item{NumServers}{integer: how many servers (1 is default)?}
}
\value{
formatted tibble
}
\description{
Outputs a formatted tibble containing the arrival time, chosen server and the service beginning and ending times of each customer, dependent on custom inputted times.
}
\examples{
Multiserver(seq(10, 100, 10), rexp(10, 1/100)) #both vectors of length 10

}
